Add files via upload
About Dataset
ICC Men's T20 World Cup Dataset
This dataset contains detailed information about the matches played in the ICC Men's T20 World Cup. It includes the following columns:

Match No.: The sequential number of the match.
Date: The date when the match was played.
Venue: The location where the match took place.
1st Team: The team that batted first.
2nd Team: The team that batted second.
Stage: The stage of the tournament (e.g., Group A, Group B, etc.).
Toss: The team that won the toss.
Winning Toss Decision: The decision made by the team winning the toss (e.g., Batting, Fielding).
First Innings Score: The score made by the team batting first.
Fall of Wickets First Innings: The number of wickets fallen in the first innings.
Second Innings Score: The score made by the team batting second.
Fall of Wickets Second Innings: The number of wickets fallen in the second innings.
Winners: The team that won the match.
Method: The method by which the match was decided (e.g., Normal Match or Duckworth Lewis Stern (DLS)).
Won by: The method of victory (e.g., Runs, Wickets).
Winning Margin: The margin of victory.
Top Scorer: The player with the highest score in the match.
Highest Score: The highest individual score made by a player in the match.
Best Bowler: The player with the best bowling performance.
Best Bowler Figure: The bowling figures of the best bowler.
Player Of The Match: The player adjudged as the best performer of the match.

This dataset provides comprehensive insights into the performances of teams and players during the ICC Men's T20 World Cup, making it a valuable resource for sports analysts and cricket enthusiasts.